+++ NEWS.md
@@ -1,13 +1,16 @@
 ## Release 0.3
 __Breaking changes__
 
-None
+* It is now ONLY possible to convert an `Option{T1}` to an `Option{T2}`, if `T1 <: T2`.
+* It is now ONLY possible to convert a `Result{O1, E1}` to a `Result{O2, E2}` if `O1 <: O2` and `E1 <: E2`.
 
 __New features__
 
 * An `Option{T}` can now be constructed from a `Result{T}`, yielding the error value if the result contains an error value.
 * A `Result{O, E}` can now be constructed using `Result(::Option{O}, e::E)`. If the option contains an error value, this will result in `Err{O, E}(e)`
 * New function `flatten`. Turns an `Option{Option{T}}` into an `Option{T}`.
+* New function: `unwrap_err`. Unwraps the wrapped error type of a `Result`, and throws an error if the `Result` conains an `Ok`
+* New function: `expect_err`. Similar to `unwrap_err`, but throws an error with a custom message.
 
 ## Release 0.2
 __Breaking changes__